# openapi.yaml
...
openapi: 3.1.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ALEXLJWOOD24/Codosseum/1.0.0
info:
  description: "This document specifies the HTTP API of a codosseum backend server."
  version: "0.1.0"
  title: Codosseum API
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
paths:
...
  /games/{id}:
    parameters:
      - $ref: "#/components/parameters/pathGameId"

    get:
      tags: [games]
      operationId: getGame
      summary: Get everything known about a game
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: game does not exist

    patch:
      tags: [games]
      operationId: updateGame
      summary: Update game settings
      description: (Partially) updates the settings of an existing game. This is only possible during the warmup phase of the game.
      security:
        - gameKey: [admin]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameSettings"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSettings"
        "400":
          description: malformed request body
        "401":
          description: No valid key provided
        "403":
          description: Key doesn't have admin privileges (cannot update game)
        "404":
          description: game does not exist
        "409":
          description: You have permission to edit this game, but it is already in progress

    delete:
      tags: [games]
      operationId: deleteGame
      summary: Delete a game
      description: Can be used by the admin of a game to delete it while in warmup.
      security:
        - gameKey: [admin]
      responses:
        "204":
          description: Sucessfully deleted
        "401":
          description: No valid key provided
        "403":
          description: Key doesn't have admin privileges
        "404":
          description: game does not exist
        "409":
          description: You have permission to delete this game, but it is already in progress
  /games/{id}/players:
    parameters:
      - $ref: "#/components/parameters/pathGameId"

    post:
      tags: [games]
      operationId: joinGame
      summary: Join a game

      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Player"

      responses:
        "200":
          description: Successfully joined the game
          content:
            application/json:
              schema:
                type: object
                required:
                  - key
                properties:
                  key:
                    type: string
                    description: A submission key that can be used to submit code to the game
        "400":
          description: Malformed request body
        "404":
          description: If the game does not exist
        "409":
          description: |
            If the game is not in a joinable state,
            or if the player has already joined the game (by IP address),
            or if a username is provided and it is already taken
    get:
      tags: [games]
      operationId: getPlayers
      summary: Get list of players in a game
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Players"
        "404":
          description: The game does not exist
  /games/{id}/players/@self:
    parameters:
      - $ref: "#/components/parameters/pathGameId"

    delete:
      tags: [games]
      operationId: leaveGame
      summary: Leave a game
      security:
        - gameKey: []
      responses:
        "204":
          description: Successfully left game
        "401":
          description: Not authenticated with a submission key
        "409":
          description: Not a player of this game (anymore)
        "404":
          description: If the game does not exist
...
  /games/{id}/solutions/test:
    parameters:
      - $ref: "#/components/parameters/pathGameId"

    post:
      tags: [games, solutions]
      operationId: testSolution
      summary: Test your solution
      description: Test a given solution to a game's challenges, returning the test results without submitting the solution
      security:
        - gameKey: []

      parameters:
        - name: testNumbers
          in: query
          description: The test numbers (public tests) to run the code against. If omitted, all tests will be run.
          schema:
            type: array
            items:
              type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Submission"

      responses:
        "200":
          # Should we introduce a mechanism to cancel tests?
          description: Acknowledgement that the solution was received and is being tested, results will be sent via Server-Sent Events
          content:
            application/json:
              schema:
                type: object
                properties:
                  testId:
                    type: string
                    description: An string to allow clients to associate future incoming test results with which request they respond to.

        "400":
          description: malformed request body

        "404":
          description: game does not exist

        "409":
          description:
            If the game is not in a state where solutions can be tested,
            or the player has already submitted a solution,
            or the player is not in the game
  /games/{id}/solutions/submit:
    parameters:
      - $ref: "#/components/parameters/pathGameId"
    post:
      tags: [games, solutions]
      operationId: submitSolution
      summary: Submit a solution
      description: Submit a solution to a game's challenges
      security:
        - gameKey: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Submission"

      responses:
        "204":
          description: Acknowledgement that the solution was received and is being tested, results will be sent via Server-Sent Events. The submission token is invalidated and may not be used again.

        "400":
          description: malformed request body

        "404":
          description: game does not exist

        "409":
          description:
            If the game is not in a state where solutions can be submitted,
            or the player has already submitted a solution,
            or the player is not in the game
...
components:
  securitySchemes:
    gameKey:
      type: apiKey
      description: |
        A key belonging to a player of a game.
        Is used to identify players when they perform actions relating to themselves,
        such as submitting a solution or leaving their game.

        Game keys can be either *admin* (role "admin") or *regular* keys (no roles).
        An admin key is given to the creator of a game and enables management of the game.
      name: Authorization
      in: header
  parameters:
    pathGameId:
      in: path
      name: id
      required: true
      description: Game ID (returned from creating a game)
      schema:
        type: string
...
  schemas:
...
    Submission:
      type: object
      required:
        - language
        - code
      properties:

        language:
          type: string
          description: The language of the solution attempt, as an identifier supported by Piston

        code:
          type: string
          description: The code to test, which is assumed to be in the language specified
...
    PlayerGameResult:
      type: object
      description: The final result of a player, after all rounds (and therefore the game) are over.
      properties:
        score:
          type: number
          description: |
            An overall (abstract) score determined by the game mode.
            This score is computed based on the cumulative performance in the rounds in which the player participated.
            It is abstract in the sense that every game mode computes it differently – It's only transparent that higher
            equals better.
...
    Game:
      type: object
      description: Information about a game
      required: [players, state, gameMode]
      properties:
        settings:
          $ref: "#/components/schemas/GameSettings"
          required: [allowedGameModes, allowedLanguages, maxPlayers, timeLimit, maxWarmupTime]
        gameMode:
          $ref: "#/components/schemas/GameMode"
          description: Active game mode
        id:
          type: string
          description: Game ID
        state:
          $ref: "#/components/schemas/GameState"
          description: current game state
        timeLeft:
          type: integer
          description: Time left on the clock for the current game state. If state is "gameOver", this value is absent
        nextStateAt:
          type: integer
          description: Unix (seconds) timestamp for when the current timer runs out. Alternative to timeLeft.
        players:
          $ref: "#/components/schemas/Players"
        results:
          type: array
          description: |
            Game results. This is different from *round results* in that it combines the results from all rounds thus far
            and interprets them according to the game mode being played. Progressive/partial results are given throughout the game,
            even if it is not over yet. Ordered from best (so far) to worst.
          items:
            $ref: "#/components/schemas/PlayerGameResult"
        round:
          type: integer
          description: Number of  the currently active round ("inProgress") or the most recent round ("roundOver", "gameOver"). Absent if state = "warmup"
    Player:
      type: object
      description: Player object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 36
    Players:
      type: object
      description: Information about the players in a game

      required:
        - players
        - admin

      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
          description: The names of the players currently in the game

        admin:
          type: string
          description: The name of the player who created the game
    GameState:
      type: string
      enum: ["warmup", "inProgress", "roundOver", "gameOver"]
      description: The available game states
    GameMode:
      type: string
      enum: ["fastest", "reverse", "golf", "arcade", "battleRoyale"]
      description: A game mode supported by codosseum
    GameSettings:
      type: object
      description: Settings for a Codosseum game

      properties:
        allowedLanguages:
          type: array
          items:
            type: string
          description: Which programming languages are allowed to be used for submissions. If omitted, all languages are allowed.
          minItems: 1

        allowedGameModes:
          description: Game modes from which the server will make a random selection.
          type: array
          items:
            $ref: "#/components/schemas/GameMode"

        maxPlayers:
          type: integer
          minimum: 2
          default: 10
          maximum: 50
          description: Maximum player count for the game

        timeLimit:
          type: integer
          minimum: 60
          maximum: 1800
          default: 900
          description: Time limit for the game, in seconds.

        maxWarmupTime:
          type: integer
          default: 60
          minimum: 0
          maximum: 300
          description: Maximum warmup time for a game in seconds - that is, a timer that starts once 2 or more players have joined the game that delays the starting of the game to allow more players to join.